+++ pyffi/spells/nif/modify.py
@@ -353,6 +353,7 @@ class SpellReverseAnimation(NifSpell):
         else:
             # recurse further
             return True
+
 class SpellCollisionMaterial(NifSpell):
     """Sets the object's collision material to be a different type"""
 
@@ -445,6 +446,35 @@ class SpellDelVertexColorProperty(NifSpell):
         # recurse further
         return True
 
+class SpellDelBlocks(NifSpell):
+    """Delete blocks that match the exclude list."""
+
+    SPELLNAME = "modify_delblocks"
+    READONLY = False
+
+    def _branchinspect(self, branch):
+        """This spell inspects every branch, also the non-admissible ones,
+        therefore we must override this method.
+        """
+        return True
+
+    def branchentry(self, branch):
+        """Strip branch if it is admissible (as specified by include/exclude
+        options of the toaster).
+        """
+        # check if it is excluded or not
+        if not self.toaster.isadmissiblebranchtype(branch.__class__):
+            # it is, wipe it out
+            self.toaster.msg("stripping this branch")
+            self.data.replace_global_node(branch, None)
+            # do not recurse further
+            return False
+        else:
+            # this one was not excluded, keep recursing
+            return True
+
+# identical to niftoaster.py modify_delblocks -x NiAlphaProperty
+# delete?
 class SpellDelAlphaProperty(NifSpell):
     """Delete alpha property if it is present."""
 
@@ -470,6 +500,8 @@ class SpellDelAlphaProperty(NifSpell):
         # recurse further
         return True
 
+# identical to niftoaster.py modify_delblocks -x NiSpecularProperty
+# delete?
 class SpellDelSpecularProperty(NifSpell):
     """Delete specular property if it is present."""
 
@@ -495,6 +527,8 @@ class SpellDelSpecularProperty(NifSpell):
         # recurse further
         return True
 
+# identical to niftoaster.py modify_delblocks -x NiBSXFlags
+# delete?
 class SpellDelBSXextradatas(NifSpell):
     """Delete BSXflags if any are present."""
 
@@ -520,6 +554,8 @@ class SpellDelBSXextradatas(NifSpell):
         # recurse further
         return True
 		
+# identical to niftoaster.py modify_delblocks -x NiStringExtraData
+# delete?
 class SpellDelNiStringExtraDatas(NifSpell):
     """Delete NiSringExtraDatas if they are present."""